requirements.txt
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
httpx>=0.26.0
Pillow>=10.0.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
elevenlabs>=1.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
